bluewow-activation/bluewow-activation/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
pydantic==2.6.4
openpyxl==3.1.5
apscheduler==3.10.4
larksuite-oapi==2.0.7
PyJWT==2.9.0
